requirements.txt
markitdown==0.0.1a2
streamlit==1.29.0
openai==1.0.0
httpx==0.24.1
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
flake8==6.1.0
